backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
flask-jwt-extended==4.7.1
bcrypt==4.3.0
pymysql==1.1.0
cryptography==41.0.4
python-dateutil==2.9.0
scikit-learn==1.3.2
joblib==1.3.2
